+++ app/routes/components/Toast/changelog.codegen.ts
@@ -8,4 +8,5 @@ export const changelog = ''
   + '- 2025-10-05: Code formatting improvement.\n'
   + '- 2025-10-28: Use `Modal.Raw` to render the toast container.\n'
   + '- 2025-11-19: Remove the dependency on `ThemeProvider`.\n'
+  + '- 2025-11-28: Enhance this with more toast types and optional description.\n'
   ;