+++ src/components/Toast/toast.tsx
@@ -1,15 +1,19 @@
 import classNames from "classnames";
-import type { FC } from "react";
+import {
+  CircleCheckIcon,
+  InfoIcon,
+  OctagonXIcon,
+  TriangleAlertIcon,
+} from "lucide-react";
+import { type FC, useMemo } from "react";
 import { useStore } from "zustand";
-
 import { useJss, useTheme } from "@/components/CopyUiProvider";
 import { Modal } from "@/components/Modal";
-
 import styles from "./toast.module.scss";
 import { ToastContext, useToastStore } from "./toast-context";
+import type { Toast as TypeToast } from "./toast-store";
 
 const ToastContainer: FC = () => {
-  const theme = useTheme();
   const jss = useJss();
 
   const store = useToastStore();
@@ -17,9 +21,7 @@ const ToastContainer: FC = () => {
   const toasts = useStore(store, (state) => state.toasts);
 
   const stx = jss.hash({
-    "--toast-bg-color": "white",
-    "--toast-border-color": theme.colors.gray["300"],
-    "--toast-box-shadow": "0 2px 4px rgba(0, 0, 0, 0.05)",
+    "--toast-boxShadow": "0 2px 4px rgba(0, 0, 0, 0.05)",
   });
 
   return (
@@ -27,14 +29,7 @@ const ToastContainer: FC = () => {
       <section className={classNames(styles.toastContainer, stx)}>
         <ul className={styles.toastList}>
           {toasts.map((toast) => (
-            <li
-              key={toast.id}
-              className={styles.toast}
-              onMouseEnter={() => store.getState().hoverToast?.(toast.id)}
-              onMouseLeave={() => store.getState().unhoverToast?.(toast.id)}
-            >
-              {toast.content}
-            </li>
+            <ToastInternal key={toast.id} toast={toast} />
           ))}
         </ul>
       </section>
@@ -44,13 +39,83 @@ const ToastContainer: FC = () => {
 
 ToastContainer.displayName = "Toast.Container";
 
-export function useToast() {
+type ToastInternalProps = {
+  toast: TypeToast;
+};
+
+const ToastInternal: FC<ToastInternalProps> = ({ toast }) => {
+  const jss = useJss();
+  const theme = useTheme();
   const store = useToastStore();
 
-  return {
-    addToast: store.getState().addToast,
-  };
-}
+  const icon = useMemo(() => {
+    switch (toast.content.type) {
+      case "success":
+        return <CircleCheckIcon />;
+      case "info":
+        return <InfoIcon />;
+      case "warning":
+        return <TriangleAlertIcon />;
+      case "error":
+        return <OctagonXIcon />;
+    }
+    return null;
+  }, [toast.content.type]);
+
+  const color = useMemo(() => {
+    switch (toast.content.type) {
+      case "success":
+        return "green";
+      case "info":
+        return "gray";
+      case "warning":
+        return "yellow";
+      case "error":
+        return "red";
+    }
+    return "gray";
+  }, [toast.content.type]);
+  const stx = jss.hash({
+    "--toast-bgColor":
+      color === "gray"
+        ? "white"
+        : `color-mix(in srgb, ${theme.colors[color]["000"]}, white)`,
+    "--toast-bdColor":
+      color === "gray"
+        ? theme.colors.gray["300"]
+        : `color-mix(in srgb, ${theme.colors[color]["300"]}, ${theme.colors.gray["300"]})`,
+    "--toast-iconColor":
+      color === "gray"
+        ? theme.colors.gray["800"]
+        : `color-mix(in srgb, ${theme.colors[color]["800"]}, ${theme.colors.gray["900"]} 25%)`,
+    "--toast-titleColor":
+      color === "gray"
+        ? theme.colors.gray["800"]
+        : `color-mix(in srgb, ${theme.colors[color]["800"]}, ${theme.colors.gray["900"]} 25%)`,
+    "--toast-descriptionColor":
+      color === "gray"
+        ? theme.colors.gray["600"]
+        : `color-mix(in srgb, ${theme.colors[color]["500"]}, ${theme.colors.gray["800"]} 25%)`,
+  });
+
+  return (
+    <li
+      className={classNames(styles.toast, icon != null && styles.withIcon, stx)}
+      onMouseEnter={() => store.getState().hoverToast?.(toast.id)}
+      onMouseLeave={() => store.getState().unhoverToast?.(toast.id)}
+    >
+      {icon != null && <div className={styles.toastIcon}>{icon}</div>}
+      <div className={styles.toastMessage}>
+        <div className={styles.toastMessageTitle}>{toast.content.message}</div>
+        {toast.content.description != null && (
+          <div className={styles.toastMessageDescription}>
+            {toast.content.description}
+          </div>
+        )}
+      </div>
+    </li>
+  );
+};
 
 type ToastComponent = {
   Container: typeof ToastContainer;