+++ src/components/Toast/CHANGELOG.md
@@ -4,3 +4,4 @@
 - 2025-10-05: Code formatting improvement.
 - 2025-10-28: Use `Modal.Raw` to render the toast container.
 - 2025-11-19: Remove the dependency on `ThemeProvider`.
+- 2025-11-28: Enhance this with more toast types and optional description.